Support onion management and fall back for self-hosted origins
