+++ lib/agent-host-service.js
@@ -5,10 +5,10 @@ import { randomUUID } from 'node:crypto';
 import { unlink, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import {
-  AGENT_HOST_CAPABILITY_LIST, AGENT_HOST_PROTOCOL_VERSION,
+  AGENT_HOST_CAPABILITY_LIST, AGENT_HOST_PROTOCOL_VERSION, isAllowedCompanionManagementParent,
 } from '../shared/agent-host-protocol.js';
 import { startExternalAgentTurn } from './agent-host-external-turn.js';
-import { createCompanionManagement, isAllowedCompanionManagementParent } from './companion-management.js';
+import { createCompanionManagement } from './companion-management.js';
 import {
   AGENT_BASE_INSTRUCTIONS, DEFAULT_TOOL_TIMEOUT_MS, DISCOVERY_SESSION_TTL_MS,
   IMAGE_EXTENSIONS, MAX_DISCOVERY_SESSIONS, MAX_IMAGE_BYTES, MAX_IMAGES_PER_TURN,
