+++ lib/companion-management.js
@@ -2,20 +2,6 @@
 // Local management UI, optionally framed by an approved getbased origin. Never exports installation secrets.
 import { randomUUID } from 'node:crypto';
 
-// Management has a narrower trust boundary than chat: custom chat origins and
-// arbitrary loopback ports must never receive privileged frame access.
-const MANAGEMENT_PARENT_ORIGINS = new Set([
-  'https://getbased.health', 'https://www.getbased.health',
-  'https://app.getbased.health', 'https://beta.getbased.health',
-  'https://get-based.vercel.app', 'https://get-based-managed-subscription-v2.vercel.app',
-  'http://127.0.0.1:8000', 'http://localhost:8000',
-]);
-
-/** @param {string} origin */
-export function isAllowedCompanionManagementParent(origin) {
-  return MANAGEMENT_PARENT_ORIGINS.has(origin);
-}
-
 /** @param {{status: () => any, control: (request: Request) => Promise<Response>, now?: () => number, allowParentOrigin?: (origin: string) => boolean}} options */
 export function createCompanionManagement(options) {
   const now = options.now || Date.now;
